Moved visualization-selector
