+++ ui/packages/catalog-ui-search/src/main/webapp/component/dropdown/visualization-selector/dropdown.visualization-selector.view.js
@@ -16,7 +16,7 @@
 const Marionette = require('marionette')
 const template = require('./dropdown.visualization-selector.hbs')
 const DropdownView = require('../dropdown.view')
-import VisualizationSelector from '../../../react-component/presentation/visualization-selector/visualization-selector'
+import VisualizationSelector from '../../../react-component/visualization-selector/visualization-selector'
 import React from 'react'
 const CustomElements = require('../../../js/CustomElements.js')
